+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # CHANGELOG
 
-## 0.1.2 - 2017-07-23
+## 0.1.2 - 23/07/2017
 
+* Tag using a wrong format DD/MM/YYYY according to Keep A Changelog
 * Tag without "v" prefix
 * Add a CHANGELOG.md file for testing
 
 ## v0.1.1 - 2017-07-22
 
+* Tag using a correct format YYYY-MM-DD according to Keep A Changelog
 * Tag with a "v" prefix to check the regular expression
 * Previous version